+++ src/main/java/org/ideplugins/vale_cli_plugin/annotator/ValeConfig.java
@@ -1,0 +1,7 @@
+package org.ideplugins.vale_cli_plugin.annotator;
+
+import java.util.List;
+
+public record ValeConfig(String rootIni, List<String> configFiles,
+List<String> paths, List<String> vocab) {
+}