+++ src/test/java/org/ideplugins/vale_cli_plugin/action/ProjectSettingsToolbarActionTest.java
@@ -1,0 +1,19 @@
+package org.ideplugins.vale_cli_plugin.action;
+
+import com.intellij.testFramework.fixtures.CodeInsightTestFixture;
+import org.ideplugins.vale_cli_plugin.BaseTest;
+import org.ideplugins.vale_cli_plugin.actions.ProjectSettingsToolbarAction;
+import org.ideplugins.vale_cli_plugin.testing.PluginTest;
+import org.ideplugins.vale_cli_plugin.testing.RunInEdtExtension;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+
+@PluginTest
+@ExtendWith(RunInEdtExtension.class)
+public class ProjectSettingsToolbarActionTest extends BaseTest{
+
+    @Test
+    public void testAction(CodeInsightTestFixture codeInsightTestFixture){
+        codeInsightTestFixture.testAction(new ProjectSettingsToolbarAction());
+    }
+}